requirements.txt
apturl==0.5.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
base58==1.0.3
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.1.0
cbor==1.0.0
certifi==2020.6.20
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
constantly==23.10.4
cryptography==44.0.2
cssselect2==0.8.0
cupshelpers==1.0
Cython==0.29.28
daphne==4.1.2
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2
django-crispy-forms==2.4
djangorestframework==3.16.0
duplicity==0.8.21
ecdsa==0.18.0b1
fasteners==0.14.1
filelock==3.6.0
flatbuffers===1.12.1-git20200711.33e2d80-dfsg1-0.6
fonttools==4.57.0
future==0.18.2
git-filter-repo==2.34.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==24.7.2
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mnemonic==0.19
monotonic==1.6
more-itertools==8.10.0
msgpack==1.1.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
pillow==11.2.1
platformdirs==2.5.1
ply==3.11
protobuf==3.12.4
ptyprocess==0.7.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydyf==0.11.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==2.4.7
pyphen==0.17.2
pypng==0.0.20
PyQRCode==1.2.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
python-snappy==0.5.3
PyTrie==0.4.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.2.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
service-identity==24.2.0
six==1.16.0
sortedcontainers==2.1.0
sqlparse==0.5.3
systemd-python==234
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
u-msgpack-python==2.3.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.1.0
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
weasyprint==65.1
webencodings==0.5.1
wsaccel==0.6.3
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==7.2
zopfli==0.2.3.post1
